add long handle

+++ include/karatsuba.h
@@ -4,10 +4,10 @@
 #include <iostream>
 #include <vector>
 
-int karatsubaMul(const int a, const int b);
+long karatsubaMul(const int a, const int b);
 int giveSizeBase(int num, int base);
 void cut(const int num, const int place, int& left, int& right);
-int basePow(const int base, const int pow);
-int recursiveMul(const int x, const int y);
+long basePow(const int base, const int pow);
+long recursiveMul(const int x, const int y);
 
 #endif // KARATSUBA_H
